pyproject.toml
[project]
name = "accurecord-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Jeffrey04", email = "mycyberpet@yahoo.com" }
]
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi[standard]>=0.115.12",
    "python-dotenv>=1.1.0",
    "structlog>=25.3.0",
    "uvicorn[standard]>=0.34.2",
]

[project.scripts]
accurecord-test = "accurecord_test:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.1",
]
